% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static}
\alias{static}
\title{Static template}
\usage{
static(number_sections = FALSE, fig_width = 7, fig_height = 5,
  fig_retina = if (!fig_caption) 2, fig_caption = FALSE, dev = "png",
  toc = FALSE, toc_depth = 3, smart = TRUE, self_contained = TRUE,
  highlight = NULL, mathjax = "default", extra_dependencies = NULL,
  css = NULL, includes = NULL, keep_md = FALSE, lib_dir = NULL,
  md_extensions = NULL, pandoc_args = NULL, ...)
}
\description{
Static template
}
