% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{yaml_front_matter}
\alias{yaml_front_matter}
\title{Parse the YAML front matter from a file}
\usage{
yaml_front_matter(input, encoding = getOption("encoding"))
}
\description{
Parse the YAML front matter from a file
}
\keyword{internal}
